+++ bench/move-rb-bench/move-rb-gen-apm-queries.cpp
@@ -133,12 +133,12 @@ search_scheme_t make_scheme(int64_t k)
 struct timed_result_t { uint64_t count; uint64_t elapsed_ns; };
 
 // reads exactly n bytes from fd into buf; false on EOF or error (e.g. the peer closed its pipe end)
-static bool read_full(int fd, void* buf, size_t n)
+static bool read_full(int32_t fd, void* buf, uint64_t n)
 {
     char* p = static_cast<char*>(buf);
     while (n > 0) {
-        ssize_t r = read(fd, p, n);
-        if (r > 0) { p += r; n -= (size_t) r; }
+        int64_t r = read(fd, p, n);
+        if (r > 0) { p += r; n -= (uint64_t) r; }
         else if (r == 0) return false;     // EOF: the peer is gone
         else if (errno == EINTR) continue; // interrupted: retry
         else return false;                 // error
@@ -147,12 +147,12 @@ static bool read_full(int fd, void* buf, size_t n)
 }
 
 // writes exactly n bytes from buf to fd; false if the peer is gone (EPIPE) or on error
-static bool write_full(int fd, const void* buf, size_t n)
+static bool write_full(int32_t fd, const void* buf, uint64_t n)
 {
     const char* p = static_cast<const char*>(buf);
     while (n > 0) {
-        ssize_t w = write(fd, p, n);
-        if (w > 0) { p += w; n -= (size_t) w; }
+        int64_t w = write(fd, p, n);
+        if (w > 0) { p += w; n -= (uint64_t) w; }
         else if (w < 0 && errno == EINTR) continue;
         else return false;
     }
@@ -160,11 +160,11 @@ static bool write_full(int fd, const void* buf, size_t n)
 }
 
 // waits up to timeout_ns for fd to become readable; false on timeout or error
-static bool wait_readable(int fd, uint64_t timeout_ns)
+static bool wait_readable(int32_t fd, uint64_t timeout_ns)
 {
     struct pollfd pfd { fd, POLLIN, 0 };
-    struct timespec ts { (time_t) (timeout_ns / 1000000000ull), (long) (timeout_ns % 1000000000ull) };
-    int r;
+    struct timespec ts { (time_t) (timeout_ns / 1000000000ull), (int64_t) (timeout_ns % 1000000000ull) };
+    int32_t r;
     do { r = ppoll(&pfd, 1, &ts, nullptr); } while (r < 0 && errno == EINTR);
     return r > 0 && (pfd.revents & POLLIN);
 }
@@ -178,12 +178,12 @@ class timeout_runner {
     query_fnc_t query;
     uint64_t m;          // bytes per pattern (the child reads exactly this many per query)
     pid_t child = -1;
-    int to_child = -1;   // parent -> child: the next pattern
-    int from_child = -1; // child -> parent: the timed_result_t
+    int32_t to_child = -1;   // parent -> child: the next pattern
+    int32_t from_child = -1; // child -> parent: the timed_result_t
 
     void spawn()
     {
-        int down[2], up[2]; // down: parent->child, up: child->parent
+        int32_t down[2], up[2]; // down: parent->child, up: child->parent
         if (pipe(down) != 0 || pipe(up) != 0) { perror("pipe"); std::exit(1); }
         pid_t pid = fork();
         if (pid < 0) { perror("fork"); std::exit(1); }
@@ -212,7 +212,7 @@ class timeout_runner {
     {
         if (child <= 0) return;
         kill(child, SIGKILL);
-        int st;
+        int32_t st;
         while (waitpid(child, &st, 0) < 0 && errno == EINTR) {}
         close(to_child); close(from_child);
         child = -1; to_child = from_child = -1;
@@ -275,7 +275,7 @@ void generate()
             for (uint64_t m : ms) {
                 if (m < scheme.p) {
                     std::cerr << "skipping " << type.op << "-" << type.metric << " k" << k << " m" << m
-                              << ": pattern length < " << (int) scheme.p << " parts in the search scheme" << std::endl;
+                              << ": pattern length < " << (int32_t) scheme.p << " parts in the search scheme" << std::endl;
                     continue;
                 }
 
@@ -379,7 +379,7 @@ int main(int argc, char** argv)
     std::setlocale(LC_ALL, "C");
     std::signal(SIGPIPE, SIG_IGN);
 
-    int i = 1;
+    int32_t i = 1;
     while (i < argc && argv[i][0] == '-') {
         std::string o = argv[i++];
         auto need = [&](const char* opt) { if (i >= argc) help(std::string("error: missing argument after ") + opt); return std::string(argv[i++]); };
